+++ packages/in_app_purchase/in_app_purchase_android/CHANGELOG.md
@@ -1,8 +1,12 @@
-# 0.1.4+4
+## NEXT
+
+* Updated Android lint settings.
+
+## 0.1.4+4
 
 * Removed dependency on the `test` package.
 
-# 0.1.4+3
+## 0.1.4+3
 
 - Updated installation instructions in README.